drivers: input: atmel: make chip calibration optional
Calibrations must be performed with a clean and dried screen.
Offer the user the possibility to disable calibrations in order
to prevent miscalibrations. Chip calibrations are not required.

Disable calibrations by default to prevent issues.
An initial miscalibration will badly affect the behaviour of the
touchscreen until a new calibration is performed, which could never
happen in case the user disabled it.